Security implementation

+++ mockenger-standalone-frontend/src/main/resources/static/modules/main/controllers/RequestListController.js
@@ -4,13 +4,14 @@ angular.module('mockengerClientMainApp')
     .controller('RequestListController', [
         '$scope',
         '$filter',
+        '$confirm',
         'projectListService',
         'groupListService',
         'requestService',
         'requestListService',
         'REQUESTS_PER_PAGE',
 
-        function($scope, $filter, projectListService, groupListService, requestService, requestListService, REQUESTS_PER_PAGE) {
+        function($scope, $filter, $confirm, projectListService, groupListService, requestService, requestListService, REQUESTS_PER_PAGE) {
             $scope.filteredData = {};
             requestListService.setLimit(REQUESTS_PER_PAGE);
 
@@ -41,26 +42,30 @@ angular.module('mockengerClientMainApp')
             }
 
             $scope.deleteRequest = function(requestToDelete) {
-                var index = requestListService.getRequestIndex(requestToDelete.id);
-                if (index < 0) {
-                    $scope.showRedMessage({data: {errors: new Array('Delete process aborted. Cannot find request in the list')}});
-                } else {
-                    var paramsToSend = {
-                        projectId: projectListService.getCurrent().id,
-                        groupId: groupListService.getCurrent().id,
-                        requestId: requestToDelete.id
-                    };
-                    requestService.ajax.delete(paramsToSend, function(response, getResponseHeaders) {
-                        $scope.showGreenMessage('Mock-request <b>' + requestToDelete.name + '</b> deleted');
-                        requestListService.removeFromRequestList(index);
+                $confirm({
+                    text: "Do you really want to delete this request?"
+                }).then(function() {
+                    var index = requestListService.getRequestIndex(requestToDelete.id);
+                    if (index < 0) {
+                        $scope.showRedMessage({data: {errors: new Array('Delete process aborted. Cannot find request in the list')}});
+                    } else {
+                        var paramsToSend = {
+                            projectId: projectListService.getCurrent().id,
+                            groupId: groupListService.getCurrent().id,
+                            requestId: requestToDelete.id
+                        };
+                        requestService.ajax.delete(paramsToSend, function(response, getResponseHeaders) {
+                            $scope.showGreenMessage('Mock-request <b>' + requestToDelete.name + '</b> deleted');
+                            requestListService.removeFromRequestList(index);
 
-                        if ((requestListService.getCurrentPage() + 1) > $scope.pageCount()) {
-                            $scope.prevPage();
-                        }
-                    }, function(errorResponse) {
-                        $scope.showRedMessage(errorResponse);
-                    });
-                }
+                            if ((requestListService.getCurrentPage() + 1) > $scope.pageCount()) {
+                                $scope.prevPage();
+                            }
+                        }, function(errorResponse) {
+                            $scope.showRedMessage(errorResponse);
+                        });
+                    }
+                });
             };
 
             $scope.$watchGroup([requestListService.getListOrder, requestListService.getSearchQuery], function() {
